+++ app/src/main/java/com/gb/stopwatch/model/domain/IStopwatchStateHolder.kt
@@ -1,0 +1,12 @@
+package com.gb.stopwatch.model.domain
+
+interface IStopwatchStateHolder {
+
+    fun start()
+
+    fun pause()
+
+    fun stop()
+
+    fun getStringTimeRepresentation(): String
+}
