+++ app/src/main/java/com/gb/stopwatch/model/domain/StopwatchState.kt
@@ -1,0 +1,13 @@
+package com.gb.stopwatch.model.domain
+
+sealed class StopwatchState {
+
+    data class Paused(
+        val elapsedTime: Long
+    ) : StopwatchState()
+
+    data class Running(
+        val startTime: Long,
+        val elapsedTime: Long
+    ) : StopwatchState()
+}
