+++ app/src/main/java/com/gb/stopwatch/model/domain/StopwatchStateCalculator.kt
@@ -1,0 +1,28 @@
+package com.gb.stopwatch.model.domain
+
+
+class StopwatchStateCalculator(
+    private val timestampProvider: TimestampProvider,
+    private val elapsedTimeCalculator: ElapsedTimeCalculator,
+) {
+
+    fun calculateRunningState(oldState: StopwatchState): StopwatchState.Running =
+        when (oldState) {
+            is StopwatchState.Running -> oldState
+            is StopwatchState.Paused -> {
+                StopwatchState.Running(
+                    startTime = timestampProvider.getMilliseconds(),
+                    elapsedTime = oldState.elapsedTime
+                )
+            }
+        }
+
+    fun calculatePausedState(oldState: StopwatchState): StopwatchState.Paused =
+        when (oldState) {
+            is StopwatchState.Running -> {
+                val elapsedTime = elapsedTimeCalculator.calculate(oldState)
+                StopwatchState.Paused(elapsedTime = elapsedTime)
+            }
+            is StopwatchState.Paused -> oldState
+        }
+}
