+++ app/src/main/java/com/gb/stopwatch/model/domain/StopwatchStateHolder.kt
@@ -1,0 +1,31 @@
+package com.gb.stopwatch.model.domain
+
+class StopwatchStateHolder(
+    private val stopwatchStateCalculator: StopwatchStateCalculator,
+    private val elapsedTimeCalculator: ElapsedTimeCalculator,
+    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter
+): IStopwatchStateHolder {
+
+    var currentState: StopwatchState = StopwatchState.Paused(0)
+        private set
+
+    override fun start() {
+        currentState = stopwatchStateCalculator.calculateRunningState(currentState)
+    }
+
+    override fun pause() {
+        currentState = stopwatchStateCalculator.calculatePausedState(currentState)
+    }
+
+    override fun stop() {
+        currentState = StopwatchState.Paused(0)
+    }
+
+    override fun getStringTimeRepresentation(): String {
+        val elapsedTime = when (val currentState = currentState) {
+            is StopwatchState.Paused -> currentState.elapsedTime
+            is StopwatchState.Running -> elapsedTimeCalculator.calculate(currentState)
+        }
+        return timestampMillisecondsFormatter.format(elapsedTime)
+    }
+}
