+++ app/src/main/java/com/gb/stopwatch/model/domain/TimestampProvide.kt
@@ -1,0 +1,7 @@
+package com.gb.stopwatch.model.domain
+
+class TimestampProvide : TimestampProvider {
+        override fun getMilliseconds(): Long {
+            return System.currentTimeMillis()
+        }
+    }
