+++ app/src/main/java/com/gb/stopwatch/model/domain/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.gb.stopwatch.model.domain
+
+interface TimestampProvider {
+    fun getMilliseconds(): Long
+}
